adds use game rosters hook
